Implement creating conditionless actions during runtime without hardcoding. Main class SpireAutomator no longer includes examples of hardcoded class structure; it is empty by default. Actions must be manually created during runtime and will not have conditions. Hardcoded class structure moved to SpireAutomatorExample class.
