+++ .idea/artifacts/autoenroller_jar.xml
@@ -1,7 +1,7 @@
 <component name="ArtifactManager">
-  <artifact type="jar" name="autoenroller-fork:jar">
-    <output-path>$PROJECT_DIR$/out/artifacts/autoenroller_fork_jar</output-path>
-    <root id="archive" name="autoenroller-fork.jar">
+  <artifact type="jar" name="autoenroller:jar">
+    <output-path>$PROJECT_DIR$/out/artifacts/autoenroller_jar</output-path>
+    <root id="archive" name="autoenroller.jar">
       <element id="module-output" name="autoenroller-fork" />
       <element id="extracted-dir" path="$PROJECT_DIR$/lib/selenium-java-3.0.1.jar" path-in-jar="/" />
       <element id="extracted-dir" path="$PROJECT_DIR$/lib/selenium-chrome-driver-3.0.1.jar" path-in-jar="/" />
